Fixed Output of Junk from plans figures
Thanks @Jakkumn - omitting all lines that were all capital letters did
the trick!

I also changed how the -p pages parameter works in the command line and
allowed for single pages, as well as page ranges to be used.

Finally, I fixed a bug where the header to the csv file was rewritten if
a CSV was outputted to multiple times.